server/requirements.txt
django==3.1.7
djangorestframework==3.12.2
markdown==3.3.4
django-filter==2.4.0
Pillow==8.1.2
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
django-cors-headers==3.7.0
django-rest-swagger==2.2.0